Moved Parser back to Dolla package.
