+++ src/main/java/dolla/parser/DebtsParser.java
@@ -1,36 +1,36 @@
-package parser;
+package dolla.parser;
 
 import dolla.Ui;
-import dolla.command.AddEntryCommand;
+import dolla.command.AddDebtsCommand;
 import dolla.command.Command;
 import dolla.command.ErrorCommand;
 import dolla.command.ShowListCommand;
 
-public class EntryParser extends Parser { public EntryParser(String inputLine) {
+public class DebtsParser extends Parser {
+
+    public DebtsParser(String inputLine) {
         super(inputLine);
     }
 
     @Override
     public Command handleInput(String mode, String inputLine) {
-
-        if (commandToRun.equals("entries")) { //show entry list
+        if (commandToRun.equals("debts")) { //show debt list
             return new ShowListCommand(mode);
-        } else if (commandToRun.equals("add")) {
-            String entryType = null;
+        } else if (commandToRun.equals("owe") || commandToRun.equals("borrow")) {
+            String type = commandToRun;
+            String name = null;
             double amount = 0.0;
             try {
-                entryType = DollaParser.verifyType(inputArray[1]);
+                name = inputArray[1];
                 amount = stringToDouble(inputArray[2]);
                 description = inputArray[3];
-                splitDescTime();
             } catch (IndexOutOfBoundsException e) {
-                Ui.printInvalidEntryFormatError();
+                Ui.printInvalidDebtFormatError();
                 return new ErrorCommand();
             } catch (Exception e) {
-                return new ErrorCommand(); // If error occurs, stop the method!
+                return new ErrorCommand();
             }
-            return new AddEntryCommand(entryType, amount, description, date);
-
+            return new AddDebtsCommand(type, name, amount, description);
         } else {
             return invalidCommand();
         }
